Graphical Image Segmentation using Opencv
This is an implementation of algorithm "Efficient Graph-Based Image
Segmentation" with OpenCV. Which is available on
http://cs.brown.edu/~pff/segment/
It is based on the paper :
P. Felzenszwalb, D. Huttenlocher
Efficient Graph-Based Image Segmentation
International Journal of Computer Vision, Vol. 59, No. 2, September 2004

The original source code is required the input image is "ppm" format.  I
rewrite this algorithm using OpenCV, making it accept more image formats
as input.

It is rewritten using standard lib so it can run under windows and
linux.